requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0

argcomplete==1.8.1
arrow==1.3.0
asgiref==3.7.2
astunparse==1.6.3
bcrypt==3.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.4
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
colored==2.2.4

contourpy==1.2.0
cookiecutter==2.5.0
cryptography==3.4.8

cycler==0.12.1
dbus-python==1.2.18

Deprecated==1.2.14
distro==1.7.0

dj-rest-auth==6.0.0
Django==4.2.7
django-allauth==0.63.3
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.4
django-cors-headers==4.4.0
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

emoji==2.12.1
exceptiongroup==1.2.1
fasteners==0.14.1
flatbuffers==23.5.26
fonttools==4.49.0
future==0.18.2
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0

h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.1.3
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.5

launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
love==0.0.3
lxml==5.1.0
macaroonbakery==1.3.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
natsort==8.0.2
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.28.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
overprint==0.0.2
packaging==23.2
Panda3D==1.10.14
panda3d-gltf==1.1.0
panda3d-simplepbr==0.12.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pkgtools==0.7.3
pluggy==1.5.0
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
PyGithub==2.3.0
Pygments==2.17.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
pytest==8.2.1
pytest-rerunfailures==14.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-pam==1.8.4
python-slugify==8.0.1
python-xapp==2.2.1
python-xlib==0.29
pytz==2022.1
pyxattr==0.7.2
PyYAML==5.4.1
rcssmin==1.1.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rich==13.7.0
rjsmin==1.2.1
rsa==4.9
scipy==1.12.0
screeninfo==0.8.1
sdkmanager==0.5.1
SecretStorage==3.3.1
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
systemd-python==234
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
text-unidecode==1.3
tinycss==0.4
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.66.4
TravisPy==0.3.5
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
ursina==6.1.2
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.7.0
withlog==0.0.3
wrapt==1.14.1
wxPython==4.0.7
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
